Rename contactInformation.js
